Move frontend script code to the Quantity Selector frontend script
